+++ src/aimanager/HISTORY.rst
@@ -6,8 +6,7 @@ Release History
 1.1.0
 ++++++
 * ``az aimanager``: Add ``get-credentials`` command to retrieve the AI Manager kubeconfig.
-* ``az aimanager namespace``: Add ``get-credentials``, ``list-accesskeys`` and
-  ``rotate-accesskeys`` commands.
+* ``az aimanager namespace``: Add ``get-credentials`` command to retrieve the namespace kubeconfig.
 
 1.0.0
 ++++++